Do not check dependencies in bundles

+++ SciQLop/sciqlop_launcher.py
@@ -40,7 +40,7 @@ def check_dependency(dependency: str) -> bool:
             if operator not in operators:
                 raise ValueError(f"Invalid operator: {operator}")
             if '*' in version:
-                if matched := re.compile(re.escape(version).replace('\*', '.*')).match(_installed_version):
+                if matched := re.compile(re.escape(version).replace(r'\*', '.*')).match(_installed_version):
                     version = matched.string
                 else:
                     return False
@@ -60,17 +60,18 @@ def install_missing_dependencies():
 
 def ask_for_installation(missing_dependencies: List[str]):
     from tkinter import messagebox
-    return messagebox.askyesno("SciQLop", "Would you like to install the missing dependencies?\n\n"
-                                          f"Missing dependencies: {', '.join(missing_dependencies)}")
+    return messagebox.askyesno("SciQLop", "Would you like to install the following missing dependencies?\n\n"
+                                          f"{', '.join(missing_dependencies)}")
 
 
 def run_sciqlop():
-    if missing := missing_dependencies():
-        if ask_for_installation(missing):
-            if not install_missing_dependencies():
-                from tkinter import messagebox
-                if not messagebox.askyesno("SciQLop", "Failed to install missing dependencies. Continue anyway?"):
-                    return 1
+    if not os.environ.get('SCIQLOP_BUNDLED', False):
+        if missing := missing_dependencies():
+            if ask_for_installation(missing):
+                if not install_missing_dependencies():
+                    from tkinter import messagebox
+                    if not messagebox.askyesno("SciQLop", "Failed to install missing dependencies. Continue anyway?"):
+                        return 1
     return run([sys.executable, "-m", "SciQLop.sciqlop_app"]).returncode
